TES/requirements.txt
Flask
Flask-SocketIO
PyCryptodome
python-socketio[client]
eventlet # hoặc gevent, tùy thuộc vào cách bạn chạy SocketIO